Greggory

+++ src/Controller/DesactiverSupprimerParticipantsController.php
@@ -6,14 +6,20 @@ use App\Entity\Participant;
 use App\Repository\ParticipantRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 class DesactiverSupprimerParticipantsController extends AbstractController
 {
+
     #[Route('/admin/desactiver-supprimer-participants', name: 'app_desactiver_supprimer_participants')]
-    public function index(ParticipantRepository $participantRepository): Response
+    public function index(Security $security,ParticipantRepository $participantRepository): Response
     {
+        if (!$security->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_main');
+        }
+
         $participantslist = $participantRepository->findAll();
         return $this->render('desactiver_supprimer_participants/index.html.twig', [
             'controller_name' => 'DesactiverSupprimerParticipantsController',
@@ -22,8 +28,12 @@ class DesactiverSupprimerParticipantsController extends AbstractController
     }
 
     #[Route('/admin/desactiver-supprimer-participants/supprimer/{id}', name: "participant_supprimer")]
-    public function supprimer(EntityManagerInterface $entityManager, Participant $participant): Response
+    public function supprimer(Security $security,EntityManagerInterface $entityManager, Participant $participant): Response
     {
+        if (!$security->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_main');
+        }
+
         $entityManager->remove($participant);
         $entityManager->flush();
 
@@ -33,8 +43,12 @@ class DesactiverSupprimerParticipantsController extends AbstractController
     }
 
     #[Route('/admin/desactiver-supprimer-participants/desactiver/{id}', name: "participant_desactiver")]
-    public function desactiver(EntityManagerInterface $entityManager, Participant $participant): Response
+    public function desactiver(Security $security,EntityManagerInterface $entityManager, Participant $participant): Response
     {
+        if (!$security->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_main');
+        }
+
         $participant->setActif(false);
         $entityManager->flush();
 
@@ -44,8 +58,12 @@ class DesactiverSupprimerParticipantsController extends AbstractController
     }
 
     #[Route('/admin/desactiver-supprimer-participants/activer/{id}', name: "participant_activer")]
-    public function activer(EntityManagerInterface $entityManager, Participant $participant): Response
+    public function activer(Security $security, EntityManagerInterface $entityManager, Participant $participant): Response
     {
+        if (!$security->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_main');
+        }
+
         $participant->setActif(true);
         $entityManager->flush();
